+++ appinfo/routes.php
@@ -47,6 +47,7 @@ return ['routes' => [
 	['name' => 'musicApi#artistDetails',	'url' => '/api/artists/{artistId}/details',	'verb' => 'GET'],
 	['name' => 'musicApi#similarArtists',	'url' => '/api/artists/{artistId}/similar',	'verb' => 'GET'],
 	['name' => 'musicApi#albumDetails',		'url' => '/api/albums/{albumId}/details',	'verb' => 'GET'],
+	['name' => 'musicApi#setPlayingTrack',	'url' => '/api/tracks/{trackId}/playing',	'verb' => 'POST'],
 	['name' => 'musicApi#scrobble',			'url' => '/api/tracks/{trackId}/scrobble',	'verb' => 'POST'],
 
 	// Search API